ReactFlow: Made into a component 'canvas'
